+++ Tests/Functional/Domain/Repository/VenueRepositoryTest.php
@@ -47,6 +47,7 @@ final class VenueRepositoryTest extends FunctionalTestCase
         self::assertSame('Alex', $result->getContactPerson());
         self::assertSame('alex@example.com', $result->getEmailAddress());
         self::assertSame('+49 1234 56789', $result->getPhoneNumber());
+        self::assertSame('Markplatz 1, 12345 Bonn', $result->getFullAddress());
     }
 
     /**